Spatial reconstruction

Fill the gaps in the MODIS LST scene using GAM with 3D spatial trend surface, and two covariables used to improve the results. The covariables have to be raster objects and matching the input MODIS LST scene.

fillingInSpace(
  lstInput,
  studyarea = "ATA",
  covariable1,
  covariable2,
  nc = 4,
  bs = "cr",
  k = 20
)

Arguments

lstInput
MODIS LST scene to be filled in space, using GAM with 3D spatial surface trend.

studyarea
Name of study area ( 'GADM','countries'), the default study area is Antarctica "ATA".

covariable1
Digital elevation model as additional variable to improve the prediction of LST values. It should be a raster object and match the extent and origin of "lstInput" variable.

covariable2
Any additional variable to improve the prediction of LST values. It should be a raster object and match the extent and origin of "lstInput" variable.

nc
Cluster size (Default = 4).

bs
Represent the smooth function and quadratic penalty to control degree of freedom of smoothing (Default value is cubic spline basis "cr").

k
Number of knots used for smoothing the function and decide the degree of freedom (Default = 20).

Value

Filled MODIS LST scene.

Examples

aqua <- raster(system.file("MODIS_data","aqua.tif", package="modislst"))

#> Error in raster(system.file("MODIS_data", "aqua.tif", package = "modislst")): could not find function "raster"

dem <- raster(system.file("covariables","dem.tif", package="modislst"))

#> Error in raster(system.file("covariables", "dem.tif", package = "modislst")): could not find function "raster"

aspect <- raster(system.file("covariables","aspect.tif", package="modislst"))

#> Error in raster(system.file("covariables", "aspect.tif", package = "modislst")): could not find function "raster"

fillingInSpace(aqua, covariable1=dem, covariable1=aspect)

#> Error in fillingInSpace(aqua, covariable1 = dem, covariable1 = aspect): formal argument "covariable1" matched by multiple actual arguments